+++ lib/src/algorithms/argon2_64bit.dart
@@ -67,9 +67,6 @@ class Argon2Internal extends Argon2 {
         );
 
   @override
-  int get derivedKeyLength => hashLength;
-
-  @override
   Argon2HashDigest convert(List<int> password) {
     int i, j, k, cols;
     int pass, slice, lane;
@@ -127,7 +124,7 @@ class Argon2Internal extends Argon2 {
     //         LE32(v) || LE32(y) || LE32(length(P)) || P ||
     //         LE32(length(S)) || S ||  LE32(length(K)) || K ||
     //         LE32(length(X)) || X)
-    var blake2b = Blake2bHash(digestSize: 64);
+    var blake2b = Blake2bHash(64);
     blake2b.addUint32(lanes);
     blake2b.addUint32(hashLength);
     blake2b.addUint32(memorySizeKB);
@@ -163,7 +160,7 @@ class Argon2Internal extends Argon2 {
 
     // Take smaller hash unchanged
     if (digestSize <= 64) {
-      var blake2b = Blake2bHash(digestSize: digestSize);
+      var blake2b = Blake2bHash(digestSize);
       blake2b.addUint32(digestSize);
       blake2b.add(message);
       var hash = blake2b.digest().bytes;
@@ -176,7 +173,7 @@ class Argon2Internal extends Argon2 {
     // Otherwise, expand to digestSize by repeatedly hashing
     // and taking the first 32-bytes from the each hash
 
-    var blake2b = Blake2bHash(digestSize: 64);
+    var blake2b = Blake2bHash(64);
     blake2b.addUint32(digestSize);
     blake2b.add(message);
     var hash = blake2b.digest().bytes;