+++ lib/src/algorithms/blake2b_64bit.dart
@@ -4,6 +4,7 @@
 import 'dart:typed_data';
 
 import 'package:hashlib/src/core/block_hash.dart';
+import 'package:hashlib/src/core/mac_base.dart';
 
 /*
               | BLAKE2b          |
@@ -60,28 +61,26 @@ const _sigma = [
 ///
 /// [rfc]: https://www.ietf.org/rfc/rfc7693.html
 /// [blake2]: https://github.com/BLAKE2/BLAKE2/blob/master/ref/blake2b-ref.c
-class Blake2bHash extends BlockHashSink {
-  final Uint64List state;
-  late final Uint64List qbuffer;
-  late final List<int> _initialState;
+class Blake2bHash extends BlockHashSink with MACSinkBase {
+  final Uint64List state = Uint64List.fromList(_seed);
+  final Uint64List _initialState = Uint64List(_seed.length);
 
   @override
   final int hashLength;
 
+  late final Uint64List qbuffer = buffer.buffer.asUint64List();
+
   /// For internal use only.
-  Blake2bHash({
-    int digestSize = 64,
+  Blake2bHash(
+    int digestSize, {
     List<int>? key,
     List<int>? salt,
     List<int>? personalization,
   })  : hashLength = digestSize,
-        state = Uint64List.fromList(_seed),
         super(1024 >>> 3) {
     if (digestSize < 1 || digestSize > 64) {
       throw ArgumentError('The digest size must be between 1 and 64');
     }
-
-    qbuffer = buffer.buffer.asUint64List();
 
     // Parameter block
     state[0] ^= 0x01010000 ^ hashLength;
@@ -122,13 +121,29 @@ class Blake2bHash extends BlockHashSink {
       }
     }
 
-    _initialState = state.toList(growable: false);
+    _initialState.setAll(0, state);
   }
 
   @override
   void reset() {
     super.reset();
     state.setAll(0, _initialState);
+  }
+
+  @override
+  void init(List<int> key) {
+    if (key.length > 64) {
+      throw ArgumentError('The key should not be greater than 64 bytes');
+    }
+    reset();
+    // The first block is the key padded with zeroes
+    buffer.setAll(0, key);
+    pos = blockLength;
+    messageLength += blockLength;
+    // Parameter block
+    state[0] ^= 0x01010000 ^ hashLength ^ (key.length << 8);
+    // Save state
+    _initialState.setAll(0, state);
   }
 
   @override