+++ lib/src/blake2b.dart
@@ -2,27 +2,28 @@
 // All rights reserved. Check LICENSE file for details.
 
 import 'package:hashlib/src/algorithms/blake2b.dart';
-import 'package:hashlib/src/core/hash_base.dart';
+import 'package:hashlib/src/core/block_hash.dart';
+import 'package:hashlib/src/core/mac_base.dart';
 
 /// For generating un-keyed message digest with BLAKE2b-160.
 ///
 /// Use [Blake2b] for keyed hash generation.
-const HashBase blake2b160 = Blake2b(outputBits: 160);
+const Blake2b blake2b160 = Blake2b(160 >>> 3);
 
 /// For generating un-keyed message digest with BLAKE2b-256.
 ///
 /// Use [Blake2b] for keyed hash generation.
-const HashBase blake2b256 = Blake2b(outputBits: 256);
+const Blake2b blake2b256 = Blake2b(256 >>> 3);
 
 /// For generating un-keyed message digest with BLAKE2b-384.
 ///
 /// Use [Blake2b] for keyed hash generation.
-const HashBase blake2b384 = Blake2b(outputBits: 384);
+const Blake2b blake2b384 = Blake2b(384 >>> 3);
 
 /// For generating un-keyed message digest with BLAKE2b-512.
 ///
 /// Use [Blake2b] for keyed hash generation.
-const HashBase blake2b512 = Blake2b(outputBits: 512);
+const Blake2b blake2b512 = Blake2b(512 >>> 3);
 
 /// The BLAKE-2b is a member of BLAKE-2 family optimized for 64-bit platforms
 /// and can generate MACs efficiently.
@@ -36,8 +37,8 @@ const HashBase blake2b512 = Blake2b(outputBits: 512);
 /// This implementation is based on the [RFC-7693][rfc]
 ///
 /// [rfc]: https://www.ietf.org/rfc/rfc7693.html
-class Blake2b extends HashBase {
-  final int outputBits;
+class Blake2b extends BlockHashBase {
+  final int digestSize;
   final List<int>? key;
   final List<int>? salt;
   final List<int>? personalization;
@@ -45,77 +46,77 @@ class Blake2b extends HashBase {
   /// Creates a [Blake2b] instance to generate hash using 64-bit numbers.
   ///
   /// Parameters:
-  /// - [outputBits] The number of bits in the output. Must be a multiple of 8.
+  /// - [digestSize] The number of bytes in the output.
   /// - [key] An optional key for MAC generation. Must be less than 64 bytes.
   /// - [salt] An optional nonce. Must be exactly 16 bytes long.
   /// - [personalization] Second optional nonce. Must be exactly 16 bytes long.
-  const Blake2b({
-    this.outputBits = 512,
+  const Blake2b(
+    this.digestSize, {
     this.key,
     this.salt,
     this.personalization,
-  }) : assert(
-          (outputBits & 7) == 0,
-          'Output bits should be a multiple of 8',
-        );
+  });
 
   @override
   Blake2bHash createSink() => Blake2bHash(
-        digestSize: outputBits >>> 3,
+        digestSize,
         key: key,
         salt: salt,
         personalization: personalization,
       );
+}
 
-  /// Get a new blake2b instance generating 160-bit digest.
-  factory Blake2b.of160({
+class Blake2bMAC extends MACHashBase {
+  final int digestSize;
+  final List<int>? salt;
+  final List<int>? personalization;
+
+  /// Creates a [Blake2b] instance to generate MAC using a [key].
+  ///
+  /// Optional parameters:
+  /// - [digestSize] The number of bytes in the output.
+  /// - [salt] An optional nonce. Must be exactly 16 bytes long.
+  /// - [personalization] Second optional nonce. Must be exactly 16 bytes long.
+  const Blake2bMAC(
+    this.digestSize,
+    List<int> key, {
+    this.salt,
+    this.personalization,
+  }) : super(key);
+
+  @override
+  MACSinkBase createSink() => Blake2bHash(
+        digestSize,
+        key: key,
+        salt: salt,
+        personalization: personalization,
+      );
+}
+
+extension Blake2bFactory on Blake2b {
+  Blake2b config({
     List<int>? key,
     List<int>? salt,
     List<int>? personalization,
-  }) =>
-      Blake2b(
-        outputBits: 160,
-        key: key,
-        salt: salt,
-        personalization: personalization,
-      );
+  }) {
+    return Blake2b(
+      digestSize,
+      key: key ?? this.key,
+      salt: salt ?? this.salt,
+      personalization: personalization ?? this.personalization,
+    );
+  }
 
-  /// Get a new blake2b instance generating 256-bit digest.
-  factory Blake2b.of256({
-    List<int>? key,
+  Blake2bMAC mac(
+    List<int> key, {
     List<int>? salt,
     List<int>? personalization,
-  }) =>
-      Blake2b(
-        outputBits: 256,
-        key: key,
-        salt: salt,
-        personalization: personalization,
-      );
-
-  /// Get a new blake2b instance generating 384-bit digest.
-  factory Blake2b.of384({
-    List<int>? key,
-    List<int>? salt,
-    List<int>? personalization,
-  }) =>
-      Blake2b(
-        outputBits: 384,
-        key: key,
-        salt: salt,
-        personalization: personalization,
-      );
-
-  /// Get a new blake2b instance generating 512-bit digest.
-  factory Blake2b.of512({
-    List<int>? key,
-    List<int>? salt,
-    List<int>? personalization,
-  }) =>
-      Blake2b(
-        outputBits: 512,
-        key: key,
-        salt: salt,
-        personalization: personalization,
-      );
+  }) {
+    return Blake2bMAC(
+      digestSize,
+      key,
+      salt: salt ?? this.salt,
+      personalization: personalization ?? this.personalization,
+    );
+  }
 }