+++ lib/src/blake2s.dart
@@ -3,26 +3,27 @@
 
 import 'package:hashlib/src/algorithms/blake2s.dart';
 import 'package:hashlib/src/core/hash_base.dart';
+import 'package:hashlib/src/core/mac_base.dart';
 
 /// For generating un-keyed message digest with BLAKE2s-128.
 ///
 /// Use [Blake2s] for keyed hash generation.
-const HashBase blake2s128 = Blake2s(outputBits: 128);
+const Blake2s blake2s128 = Blake2s(128 >>> 3);
 
 /// For generating un-keyed message digest with BLAKE2s-160.
 ///
 /// Use [Blake2s] for keyed hash generation.
-const HashBase blake2s160 = Blake2s(outputBits: 160);
+const Blake2s blake2s160 = Blake2s(160 >>> 3);
 
 /// For generating un-keyed message digest with BLAKE2s-224.
 ///
 /// Use [Blake2s] for keyed hash generation.
-const HashBase blake2s224 = Blake2s(outputBits: 224);
+const Blake2s blake2s224 = Blake2s(224 >>> 3);
 
 /// For generating un-keyed message digest with BLAKE2s-256.
 ///
 /// Use [Blake2s] for keyed hash generation.
-const HashBase blake2s256 = Blake2s(outputBits: 256);
+const Blake2s blake2s256 = Blake2s(256 >>> 3);
 
 /// The BLAKE-2s is a member of BLAKE-2 family optimized for 32-bit platforms
 /// and can generate MACs efficiently.
@@ -37,7 +38,7 @@ const HashBase blake2s256 = Blake2s(outputBits: 256);
 ///
 /// [rfc]: https://www.ietf.org/rfc/rfc7693.html
 class Blake2s extends HashBase {
-  final int outputBits;
+  final int digestSize;
   final List<int>? key;
   final List<int>? salt;
   final List<int>? personalization;
@@ -45,78 +46,78 @@ class Blake2s extends HashBase {
   /// Creates a [Blake2s] instance to generate hash using 32-bit numbers.
   ///
   /// Parameters:
-  /// - [outputBits] The number of bits in the output. Must be a multiple of 8.
+  /// - [digestSize] The number of bytes in the output.
   /// - [key] An optional key for MAC generation. Must be less than 32 bytes.
   /// - [salt] An optional nonce. Must be exactly 8 bytes long.
   /// - [personalization] Second optional nonce. Must be exactly 8 bytes long.
   ///
-  const Blake2s({
-    this.outputBits = 256,
+  const Blake2s(
+    this.digestSize, {
     this.key,
     this.salt,
     this.personalization,
-  }) : assert(
-          (outputBits & 7) == 0,
-          'Output bits should be a multiple of 8',
-        );
+  });
 
   @override
   Blake2sHash createSink() => Blake2sHash(
-        digestSize: outputBits >>> 3,
+        digestSize,
         key: key,
         salt: salt,
         personalization: personalization,
       );
+}
 
-  /// Get a new blake2s instance generating 128-bit digest.
-  factory Blake2s.of128({
+class Blake2sMAC extends MACHashBase {
+  final int digestSize;
+  final List<int>? salt;
+  final List<int>? personalization;
+
+  /// Creates a [Blake2s] instance to generate MAC using a [key].
+  ///
+  /// Optional parameters:
+  /// - [digestSize] The number of bytes in the output
+  /// - [salt] An optional nonce. Must be exactly 16 bytes long.
+  /// - [personalization] Second optional nonce. Must be exactly 16 bytes long.
+  const Blake2sMAC(
+    this.digestSize,
+    List<int> key, {
+    this.salt,
+    this.personalization,
+  }) : super(key);
+
+  @override
+  MACSinkBase createSink() => Blake2sHash(
+        digestSize,
+        key: key,
+        salt: salt,
+        personalization: personalization,
+      );
+}
+
+extension Blake2sFactory on Blake2s {
+  Blake2s config({
     List<int>? key,
     List<int>? salt,
     List<int>? personalization,
-  }) =>
-      Blake2s(
-        outputBits: 128,
-        key: key,
-        salt: salt,
-        personalization: personalization,
-      );
+  }) {
+    return Blake2s(
+      digestSize,
+      key: key ?? this.key,
+      salt: salt ?? this.salt,
+      personalization: personalization ?? this.personalization,
+    );
+  }
 
-  /// Get a new blake2s instance generating 160-bit digest.
-  factory Blake2s.of160({
-    List<int>? key,
+  Blake2sMAC mac(
+    List<int> key, {
     List<int>? salt,
     List<int>? personalization,
-  }) =>
-      Blake2s(
-        outputBits: 160,
-        key: key,
-        salt: salt,
-        personalization: personalization,
-      );
-
-  /// Get a new blake2s instance generating 224-bit digest.
-  factory Blake2s.of224({
-    List<int>? key,
-    List<int>? salt,
-    List<int>? personalization,
-  }) =>
-      Blake2s(
-        outputBits: 224,
-        key: key,
-        salt: salt,
-        personalization: personalization,
-      );
-
-  /// Get a new blake2s instance generating 256-bit digest.
-  factory Blake2s.of256({
-    List<int>? key,
-    List<int>? salt,
-    List<int>? personalization,
-  }) =>
-      Blake2s(
-        outputBits: 256,
-        key: key,
-        salt: salt,
-        personalization: personalization,
-      );
+  }) {
+    return Blake2sMAC(
+      digestSize,
+      key,
+      salt: salt ?? this.salt,
+      personalization: personalization ?? this.personalization,
+    );
+  }
 }